// src/enums/product.ts
export enum EProductStatus {
  AVAILABLE = "AVAILABLE",
  PENDING = "PENDING",
  SOLD = "SOLD",
  INACTIVE = "INACTIVE",
}

export enum EProductCondition {
  NEW = "NEW",
  LIKE_NEW = "LIKE_NEW",
  GOOD = "GOOD",
  FAIR = "FAIR",
  POOR = "POOR",
}

export enum EDeliveryMethod {
  SHIPPER = "SHIPPER",
  HAND_DELIVERY = "HAND_DELIVERY",
  BOTH = "BOTH",
}